api/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.13.0b3
langchain==0.0.139
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
